requirements.txt
amqp==5.1.1
aniso8601==6.0.0
APScheduler==3.10.1
billiard==3.6.4.0
celery==5.1.2
certifi==2023.7.22
charset-normalizer==3.2.0
Click==7.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Flask==1.0.3
Flask-APScheduler==1.12.4
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
idna==3.4
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==5.3.1
MarkupSafe==1.1.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2019.1
requests==2.31.0
six==1.12.0
SQLAlchemy==1.3.6
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
Werkzeug==0.15.5
